renamed reducers and action creators for data availability
